phase_2_generator/pyproject.toml
[project]
name = "phase-2-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fal-client>=0.5.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "boto3>=1.34.0",
    "firebase-admin>=6.5.0",
    "Pillow>=10.3.0",
    "python-dotenv>=1.0.1",
]
